feat: Allow to display children of found matches (#204)
* Use different classname than bootstrap

`hide` classname is being used by bootstrap to hide elements. For us, we need to change the classname here in react-dropdown-tree-select to display everything correctly.

* Option to show children on search results

This allows for children of found matches to be displayed. With this option, one can search for the parent and then see all children and select any.

* Add docs/bundle.js

* Remove unnecessary line

* nicer code

* Code optimization

* Rename addChildrensToTree

* Update docs

* ESLint

* Remove bootstrap hide workaround

* test: Updated syntax after merging latest develop

* chore: Updated bundle after merging latest develop